+++ client/docs/embed.go
@@ -1,0 +1,8 @@
+package docs
+
+import (
+	"embed"
+)
+
+//go:embed swagger-ui/*
+var SwaggerUI embed.FS